backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.1.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.4
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.2
pep8-naming==0.12.1
Pillow==9.3.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
testfixtures==6.18.5
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
